Apply new internal header/source structure on time.{c,h}
