+++ libraries/ah_base/include/ah/internal/kqueue/time.h
@@ -1,0 +1,12 @@
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// SPDX-License-Identifier: EPL-2.0
+
+#ifndef AH_INTERNAL_KQUEUE_TIME_H_
+#define AH_INTERNAL_KQUEUE_TIME_H_
+
+#define AH_I_TIME_PLATFORM_FIELDS uint64_t _mach_absolute_time;
+
+#endif
