+++ libraries/ah_base/include/ah/internal/time.h
@@ -1,0 +1,22 @@
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// SPDX-License-Identifier: EPL-2.0
+
+#ifndef AH_INTERNAL_TIME_H_
+#define AH_INTERNAL_TIME_H_
+
+#include "../defs.h"
+
+#if AH_USE_IOCP
+#    include "iocp/time.h"
+#elif AH_USE_KQUEUE
+#    include "kqueue/time.h"
+#elif AH_USE_URING
+#    include "uring/time.h"
+#endif
+
+#define AH_I_TIME_FIELDS AH_I_TIME_PLATFORM_FIELDS
+
+#endif
