+++ libraries/ah_base/include/ah/internal/uring/time.h
@@ -1,0 +1,14 @@
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// SPDX-License-Identifier: EPL-2.0
+
+#ifndef AH_INTERNAL_URING_TIME_H_
+#define AH_INTERNAL_URING_TIME_H_
+
+#include <linux/time_types.h>
+
+#define AH_I_TIME_PLATFORM_FIELDS struct __kernel_timespec _timespec;
+
+#endif
